Move escape helpers to EscapeHelper
