+++ spec/unit/helpers/escape_helper/escape_html_spec.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-RSpec.describe Hanami::View::HTML, ".escape_html" do
+require "hanami/view/helpers/escape_helper"
+
+RSpec.describe Hanami::View::Helpers::EscapeHelper, ".escape_html" do
   def escape_html(...)
     described_class.escape_html(...)
   end
